Replace hamburger icon and relocate pin sidebar button
